71.13 G.L. 12, Atlas Manager Hiring Period - Carston Haffner
The cyclical hiring process is the process by which department heads are hired for the upcoming year. The application window for department head positions are open the first two weeks of Spring quarter, with the hope that the positions can be filled promptly allowing the incoming heads to start their training beginning the third week. Currently there are two positions that have application periods that begin earlier to allow for additional training time. These positions are the FVP and MVP, and are meant to be hired before Spring quarter even begins. We, along with Hunter Giesbrecht, the assistant manager at the Atlas, feel that hiring the incoming Atlas Manager at the same time as the FVP and MVP would be very beneficial to the transition process at the Atlas. Bryan has outlined four main points that illustrate why hiring the Atlas Manager earlier would be beneficial. 

 

Reasons to hire the manager early: 

Assuming the manager is already skilled in making coffee (which is not always a guarantee) there is still much more to learn in Atlas operations such as coordinating events, meeting with local vendors, basic responsibilities as an ASWWU Head, etc. 

As of this year, Hunter (the assistant manager) and I have been fully reorganizing and updating the financial records of the Atlas and to understand the financial processes, extensive training will need to be offered. 

The staff hiring process tends to involve 60+ applicants which takes more than the allotted time to sift through even for only a few positions. 

The Atlas hiring process also involves two phases, rather than one. First phase is to hire the assistant manager, and give them a chance to get trained as well. Both phases need at least 2 weeks to be open on the website and the assistant manager usually plays a part in hiring the rest of the team. Therefore, they can’t overlap.